moving to branch tryagain

+++ src/utils/cloudinary.js
@@ -1,43 +1,43 @@
-import { v2 as cloudinary } from 'cloudinary';
-import fs from "fs"
-import path from "path"
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
+import dotenv from "dotenv"
 
+dotenv.config() // but why import dotenv again
 
-
-
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
 const uploadOnCloudinary = async (localFilePath) => {
-    try {
-        if (!localFilePath) return null
-        cloudinary.config({
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-        api_key: process.env.CLOUDINARY_API_KEY,
-        api_secret: process.env.CLOUDINARY_API_SECRET
-    })
-        // upload file on cloudinary
-        console.log("Checking file exists:", fs.existsSync(localFilePath));
-        const response = await cloudinary.uploader.upload(localFilePath , {
-            resource_type: "auto"
-        })
-        //file has been uploaded successfully
-        console.log(`file is uploaded on clodinary , url is :  ${response.url} and the log of response is ${response}`);
-        // if (localFilePath) {
-        //     fs.unlinkSync(localFilePath)
-        //     console.log(`finally uploaded to cloudinary and deleted form local machine`);
-            
-        // }
-        return response
-    } catch (error) {
-        console.log(`cloudinary upload failed ${error}`)
-        fs.unlinkSync(localFilePath) // remove the locally saved temporay file as the uplaod operation got failed         
-        return null;
+  try {
+    if (!localFilePath) return null;
+    // upload file on cloudinary
+    const response = await cloudinary.uploader.upload(localFilePath, {
+        resource_type: "auto",
+    });
+    console.log("Checking file exists:", fs.existsSync(localFilePath));
+    //file has been uploaded successfully
+    console.log(
+      `file is uploaded on clodinary , url is :  ${response.url} and the log of response is ${response}`
+    );
+
+    
+    fs.unlinkSync(localFilePath);
+    console.log(`finally uploaded to cloudinary and deleted form local machine`);
+    return response;
+  } catch (error) {
+    console.log(`cloudinary upload failed ${error}`);
+    // fs.unlinkSync(localFilePath) // remove the locally saved temporay file as the uplaod operation got failed
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+      console.log(`Local file deleted after failed upload`);
     }
-}
+    return null;
+  }
+};
 
-
-export {uploadOnCloudinary}
-
-
-
+export { uploadOnCloudinary };
 
 // import { v2 as cloudinary } from "cloudinary";
 // import fs from "fs";
